added objects and updating to gameworld. Thinking that 3d/metal is perhaps unnecessary.

+++ ProjectOI/Vector3.swift
@@ -10,7 +10,7 @@ import Foundation
 
 struct Vector3
 {
-    var x,y,z: Float
+    var x,y,z: Double
     
     func equals(vec:Vector3)->Bool
     {
